fix bug, resolve problem with LogToFile and LogToRotatingFile unit tests
Avoir un seul objet LogSaveDAO en pattern singleton c'était pas une
bonne idée vu que les cibles des logs sont spécifique à chaque Logger
donc c'était la source d'un premier bug. Au final, la classe LogSaveDAO
n'avait que deux lignes de code utile, donc je l'ai supprimée et j'ai
mis le code dans la classe Logger.

Autre problème, on avait mal compris les annotations @Before et @After
dans les tests de jUnit, en fait les deux méthodes annotées avec ça
sont exécutées avant ou après CHAQUE test, et pas avant ou après
l'ensemble des tests, donc j'ai mis des conditions pour n'exécuter
le code qu'une fois au bon moment.

Tous les tests sont au vert ;)

+++ OrnithoLog/test/com/github/encrypt/ornitholog/logtargets/TestLogToRotatingFile.java
@@ -11,6 +11,7 @@ import java.util.List;
 
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
 import com.github.encrypt.ornitholog.LogLevel;
@@ -21,24 +22,32 @@ import com.github.encrypt.ornitholog.formatters.LevelFormatter;
 
 public class TestLogToRotatingFile {
 
+	private static boolean setupDone = false;  // because we need to setup the tests only once
+	private static int testsDone = 0;   // the number of tests done
+	
 	private Logger logger;
-	private String myPath = "/Users/patouz/Documents/TestLogToRotatingFile.txt";
+	private String myPath = "TestLogToRotatingFile.txt";
 	
 	
 	@Before
 	public void before() {
-		this.logger = LoggerFactory.getLogger("com.example.TestLogToRotatingFile");
-		logger.addTarget(new LogToRotatingFile(myPath, 1024000.0));
-		logger.setFormatter(new LevelFormatter(logger));
-		logger.setLevel(LogLevel.ERROR);
-		logger.debug("Ceci est un test d'erreur!");
-		logger.debug("Ceci est le 2e test d'erreur !");
+		if (!this.setupDone) {
+			this.logger = LoggerFactory.getLogger("com.example.TestLogToRotatingFile");
+			logger.addTarget(new LogToRotatingFile(myPath, 1024000.0));
+			logger.setFormatter(new LevelFormatter(logger));
+			logger.setLevel(LogLevel.ERROR);
+			logger.debug("Ceci est un test d'erreur!");
+			logger.debug("Ceci est le 2e test d'erreur !");
+			this.setupDone = true;
+		}
 	}
 	
 	@Test
 	public void fileExists() {
 		File file = new File(myPath);
 		assertTrue(file.exists());
+		
+		testsDone++;
 	}
 	
 	@Test
@@ -50,14 +59,19 @@ public class TestLogToRotatingFile {
 			e.printStackTrace();
 			fail("Test Failed : File doesn't exists.");
 		}
+		
+		testsDone++;
 	}
 	
 	
 	@After
 	public void clean(){
-		File file = new File(myPath);
-		if(file.exists()){
-			file.delete();
+		// we should delete the file only when all the tests are done
+		if (testsDone == 2) {
+			File file = new File(myPath);
+			if(file.exists()){
+				file.delete();
+			}
 		}
 	}
 }
